Added homepage, splitted views
