+++ lib/main.dart
@@ -1,8 +1,8 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_course/firebase_options.dart';
 import 'package:flutter_course/views/login_view.dart';
-import 'firebase_options.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -20,40 +20,18 @@ void main() {
       // is not restarted.
       primarySwatch: Colors.blue,
     ),
-    home: const LoginView(),
+    home: const HomePage(),
   ));
 }
 
-class RegisterView extends StatefulWidget {
-  const RegisterView({Key? key}) : super(key: key);
-
-  @override
-  State<RegisterView> createState() => _RegisterViewState();
-}
-
-class _RegisterViewState extends State<RegisterView> {
-  late final TextEditingController _email;
-  late final TextEditingController _password;
-
-  @override
-  void initState() {
-    _email = TextEditingController();
-    _password = TextEditingController();
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _email.dispose();
-    _password.dispose();
-    super.dispose();
-  }
+class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("Register"),
+        title: const Text("Home"),
       ),
       body: FutureBuilder(
           future: Firebase.initializeApp(
@@ -62,47 +40,14 @@ class _RegisterViewState extends State<RegisterView> {
           builder: (context, snapshot) {
             switch (snapshot.connectionState) {
               case ConnectionState.done:
-                return Column(
-                  children: [
-                    TextField(
-                      controller: _email,
-                      enableSuggestions: false,
-                      autocorrect: false,
-                      keyboardType: TextInputType.emailAddress,
-                      decoration: const InputDecoration(
-                          hintText: "Enter your email here"),
-                    ),
-                    TextField(
-                      controller: _password,
-                      obscureText: true,
-                      enableSuggestions: false,
-                      autocorrect: false,
-                      decoration: const InputDecoration(
-                        hintText: "Enter your password here",
-                      ),
-                    ),
-                    TextButton(
-                        onPressed: () async {
-                          final email = _email.text;
-                          final password = _password.text;
-                          try {
-                            final userCredential = await FirebaseAuth.instance
-                                .createUserWithEmailAndPassword(
-                                    email: email, password: password);
-                            print(userCredential);
-                          } on FirebaseAuthException catch (e) {
-                            if (e.code == 'weak-password') {
-                              print("Weak password");
-                            } else if (e.code == 'email-already-in-use') {
-                              print("Email already in use");
-                            } else if (e.code == 'invalid-email') {
-                              print("Invalid Email entered");
-                            }
-                          }
-                        },
-                        child: const Text('Register')),
-                  ],
-                );
+                final user = FirebaseAuth.instance.currentUser;
+                final emailVerified = user?.emailVerified ?? false;
+                if (emailVerified) {
+                  print('You are a verified user');
+                } else {
+                  print('You need to verify your email first');
+                }
+                return const Text('Done');
               default:
                 return const Text("Loading...");
             }
